+++ pages/composition.tsx
@@ -24,36 +24,41 @@ const Composition = (): JSX.Element => {
     }
   }, [wrapperRef, groups]);
 
-  const participantWidth = (participants: number): string => {
-    let width = `${size.width / groups.length}px`;
-    if (participants >= 2) {
-      width = `${size.width / groups.length / 2}px`;
-    } else {
-      width = `${size.width / groups.length}px`;
+  const participantWidth = (participants: number, group: Group): string => {
+    let width = `${groupWidth(group)}px`;
+    let divider = 2;
+    for (let i = 2; i <= participants; i = i * 2) {
+      width = `${groupWidth(group) / divider}px`;
+      divider++;
     }
     return width;
   };
 
-  const participantHeight = (participants: number): string => {
-    let height = '100%';
-    if (participants >= 3) {
-      height = `${groupHeight() / Math.round(participants / 2)}px`;
-    } else {
-      height = '100%';
+  const participantHeight = (participants: number, mobile?: boolean): string => {
+    const mobileHeight = !mobile ? size.height : size.height / groups.length;
+    let height = `${mobileHeight}px`;
+    let divider = 2;
+    let step = 1;
+    for (let i = 3; i <= participants; i = i + step) {
+      height = `${mobileHeight / divider}px`;
+      divider++;
+      step += 3;
     }
     return height;
   };
 
-  const groupHeight = (): number => {
-    let height = size.height;
-    const participantLength: Array<number> = groups.map((group) => group.participants.length);
-    const isEqual: boolean = participantLength.every((val, i, arr) => val === arr[0]);
-    if (isEqual) {
-      height = size.height / groups.length;
-    } else {
-      //TODO calculate height if gorups have diff number of participants
-    }
-    return height;
+  const groupWidth = (group: Group) => {
+    let width: number;
+    const moreThenTen = groups.filter((group) => group.participants.length >= 10).length;
+    if (moreThenTen <= 0 || moreThenTen === groups.length) width = size.width / groups.length;
+    else if (group.participants.length >= 10)
+      width = size.width / groups.length + (size.width / groups.length) * 0.1;
+    else
+      width =
+        size.width / groups.length -
+        (size.width / groups.length) * (moreThenTen * (0.1 / (groups.length - moreThenTen)));
+
+    return width;
   };
 
   return (
@@ -63,10 +68,8 @@ const Composition = (): JSX.Element => {
           <Flex
             key={index}
             sx={{
-              width: [size.width, `${size.width / groups.length}px`],
-              minWidth: [size.width, `${size.width / groups.length}px`],
-              maxWidth: [size.width, `${size.width / groups.length}px`],
-              height: [`${groupHeight()}px`, '100vh'],
+              width: ['100vh', `${groupWidth(group)}px`],
+              height: [`${size.height / groups.length}px`, '100vh'],
               // border: '1px solid red',
             }}
           >
@@ -81,18 +84,29 @@ const Composition = (): JSX.Element => {
                   <Flex
                     key={index}
                     sx={{
-                      width: participantWidth(group.participants.length),
-                      minWidth: participantWidth(group.participants.length),
-                      height: participantHeight(group.participants.length),
-                      border: '1px solid green',
+                      width: [
+                        participantWidth(group.participants.length, group),
+                        participantWidth(group.participants.length, group),
+                      ],
+                      height: [
+                        participantHeight(group.participants.length, true),
+                        participantHeight(group.participants.length, false),
+                      ],
+                      // border: '1px solid blue'
                     }}
                   >
                     {participant.hasVideo && (
                       <VideoPlayer
                         sx={{
                           position: 'absolute',
-                          width: participantWidth(group.participants.length),
-                          height: participantHeight(group.participants.length),
+                          width: [
+                            participantWidth(group.participants.length, group),
+                            participantWidth(group.participants.length, group),
+                          ],
+                          height: [
+                            participantHeight(group.participants.length, true),
+                            participantHeight(group.participants.length, false),
+                          ],
                         }}
                         video={participant.video}
                       />